Tweaked app install script

+++ bin/applications.sh
@@ -2,33 +2,50 @@
 
 # Install development apps and platforms using Homebrew
 
-# Ask for the administrator password upfront.
-sudo -v
+echo "This script will install the following:"
+echo "APPLICATIONS"
+echo "IntelliJ IDEA, Sublime Text 3, iTerm2"
+echo ""
+echo "PLATFORMS:"
+echo "Groovy"
+echo ""
+echo "TOOLS"
+echo "xCode command line tools, Homebrew, Brew Cask"
+echo ""
 
-# Keep-alive: update existing `sudo` time stamp until the script has finished.
-while true; do sudo -n true; sleep 60; kill -0 "$$" || exit; done 2>/dev/null &
+read -p "Do you wish to continue? (y/n) " -n 1;
+	echo "";
+if [[ $REPLY =~ ^[Yy]$ ]]; then
 
-# First, install OS X commandline tools if they aren't already installed:
-xcode-select –install
+	# Ask for the administrator password upfront.
+	sudo -v
 
-# Install Homebrew if it isn't already installed:
-ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
+	# Keep-alive: update existing `sudo` time stamp until the script has finished.
+	while true; do sudo -n true; sleep 60; kill -0 "$$" || exit; done 2>/dev/null &
 
-# Install Brew Cask:
-brew install caskroom/cask/brew-cask
-brew tap caskroom/cask
+	# First, install OS X commandline tools if they aren't already installed:
+	xcode-select –install
 
-# Install Sublime Text:
-brew cask install sublime-text3
+	# Install Homebrew if it isn't already installed:
+	ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
 
-# Install IntelliJ IDEA:
-brew cask install intellij-idea
+	# Install Brew Cask:
+	brew install caskroom/cask/brew-cask
+	brew tap caskroom/cask
 
-# Install iTerm2:
-brew cask install iterm2
+	# Install Sublime Text:
+	brew cask install sublime-text3
 
-# Install Groovy:
-brew install groovy
+	# Install IntelliJ IDEA:
+	brew cask install intellij-idea
 
-# Remove outdated versions from the cellar.
-brew cleanup
+	# Install iTerm2:
+	brew cask install iterm2
+
+	# Install Groovy:
+	brew install groovy
+
+	# Remove outdated versions from the cellar.
+	brew cleanup
+
+fi;
